Returned skipped test return value 77 for testcase not relevant for servers other then netconfd. Fixed missing module dependency on iana-if-type.yang.

+++ netconf/test/netconfd/validate-config-only/run.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+
+if [ "$RUN_WITH_CONFD" != "" ] ; then
+    # skipped test return value
+    exit 77
+fi
+
 rm -rf tmp || true
 mkdir tmp
